Rename view module to viewer module
and move neurom/viewer.py to neurom/viewer/__init__.py
so that public function are accessed the same way:
```python
from neurom.viewer import draw
```

Rationale:
As of today the structure is confusing since public functions can potentially
in one of the 3 following files.
- viewer.py
- view/__init__.py
- view/view.py

Now at least viewer.py is gone and public functions are exposed in the
viewer module though the __init__.py
