+++ examples/plot_features.py
@@ -35,11 +35,11 @@ import json
 import argparse
 
 import numpy as np
-import neurom as nm
-from neurom.view import common as view_utils
 import scipy.stats as _st
 from matplotlib.backends.backend_pdf import PdfPages
 
+import neurom as nm
+from neurom.viewer import common as view_utils
 
 DISTS = {
     'normal': lambda p, bins: _st.norm.pdf(bins, p['mu'], p['sigma']),
