+++ neurom/viewer/_dendrogram.py
