+++ neurom/viewer/common.py
